move bid selection component to bids file
